+++ workflow/scripts/validation/utils.py
@@ -254,8 +254,8 @@ def transform_coordinates(
     Return a copy of the tree sequence `ts` with the coordinate system
     monotonically transformed according to `ratemap.get_cumulative_mass(x)`.
     Zero length edges are removed, and any nodes that are then disconnected are
-    removed as well.  All sites and mutations that are within zero-rate regions
-    are removed.
+    removed as well.  All sites and mutations that are within zero- or NaN-rate
+    regions are removed.
     """
     assert ratemap.sequence_length == ts.sequence_length
     tab = ts.dump_tables()
@@ -272,7 +272,8 @@ def transform_coordinates(
     tab.edges.parent = node_map[tab.edges.parent]
     tab.edges.child = node_map[tab.edges.child]
     # map sites to new coordinate system and remove those in masked intervals
-    site_map = tab.sites.keep_rows(ratemap.get_rate(tab.sites.position).astype(bool))
+    site_rate = ratemap.get_rate(tab.sites.position)
+    site_map = tab.sites.keep_rows(np.nan_to_num(site_rate).astype(bool))
     tab.sites.position = ratemap.get_cumulative_mass(tab.sites.position)
     # update mutation pointers and remove those without a node or site
     tab.mutations.node = node_map[tab.mutations.node]